requirements.txt
Pillow==2.7.0
pyFirmata==0.9.5
pyserial==2.7
wsgiref==0.1.2
